fix files.py
