+++ .github/workflows/gcp_deploy.yml
@@ -1,0 +1,45 @@
+# 質問事項
+# pytestとファイルを分ける方針で良いか？
+    # featureからmainにマージしてpytestが走って成功した後でデプロイ？
+
+
+name: Deploy to Cloud Run from main branch
+
+on:
+  push:
+    branches:
+      - main
+      
+env:
+    env:
+    GCP_PROJECT_ID: ${{ secrets.GCP_PROJECT_ID }}
+    SERVICE: ""
+    REGION: "asia-northeast1"
+    GCP_DEPLOY_SERVICE_ACCOUNT_KEY: ${{ secrets.GCP_DEPLOY_SERVICE_ACCOUNT_KEY }}  
+    BUCKET_NAME: ${{ secrets.BUCKET_NAME }}
+
+# deployはfrontとbackend別々で記載する（jobs配下でdeploy-backendとdeploy-frontendで分けてデプロイ）
+jobs:
+    deploy:
+        runs-on: ubuntu-latest
+
+        steps:
+        - uses: actions/checkout@v4
+
+        - name: backendイメージをビルドしてプッシュ
+        run: |
+          cd backend
+          gcloud builds submit --tag gcr.io/${{ secrets.GCP_PROJECT_ID }}/backend
+        
+        - name: backendイメージをビルドしてプッシュ
+        run: |
+          cd frontend
+          gcloud builds submit --tag gcr.io/${{ secrets.GCP_PROJECT_ID }}/frontend
+          
+        - name: BackendをArtifact Registryからデプロイ
+        run: |
+          gcloud run deploy backend --image gcr.io/${{ secrets.GCP_PROJECT_ID }}-registry-ai-notebook/backend --platform managed --region asia-northeast1
+ 
+        - name: frontendをArtifact Registryからデプロイ
+        run: |
+          gcloud run deploy frontend --image gcr.io/${{ secrets.GCP_PROJECT_ID }}-registry-ai-notebook/frontend --platform managed --region asia-northeast1
